+++ src/main/java/nextstep/ladder/domain/LineIF.java
@@ -1,0 +1,6 @@
+package nextstep.ladder.domain;
+
+interface LineIF {
+
+    Position moveOn(Position from);
+}
